Created new type specifically for move stats, and more..

+++ options.h
@@ -3,7 +3,8 @@
  * 
  * the MicroChess project: https://github.com/ripred/MicroChess
  * 
- * header file for MicroChess
+ * The options_t structure holds flags that are used during the game
+ * to enable or disable various features of the chess engine.
  * 
  */
 #ifndef OPTIONS_INCL
@@ -16,7 +17,7 @@ public:
     uint8_t     random : 1,     // add randomness to the game?
              profiling : 1;     // are we profiling the engine?
     uint8_t     maxply;         // the maximum ply level
-    uint8_t     move_limit;     // the maximum number of moves allowed in a full game
+    uint32_t    move_limit;     // the maximum number of moves allowed in a full game
     uint32_t    seed;           // the starting seed hash for prn's
     print_t     print_level;    // the verbosity setting for the level of output
 
@@ -37,4 +38,4 @@ public:
 
 };  // options_t
 
-#endif
+#endif  // OPTIONS_INCL
